+++ package.json
@@ -22,12 +22,13 @@
     "clean-tp-data": "node ./tools/clean.js userData TrackingProtection.dat",
     "download-languages": "node ./tools/downloadLanguages",
     "download-ethwallet": "node ./tools/downloadEthwallet",
+    "download-geth": "node ./tools/downloadGeth",
     "download-sync-client": "node ./tools/downloadSyncClient",
     "electron-rebuild": "electron-rebuild",
     "flow": "flow; test $? -eq 0 -o $? -eq 2",
     "lint": "standard --verbose | snazzy",
     "lint-standard": "standard",
-    "postinstall": "npm run download-sync-client && npm run download-ethwallet && webpack",
+    "postinstall": "npm run download-sync-client && npm run download-ethwallet && npm run download-geth && webpack",
     "pre-push-tests": "if [ -z $NO_PUSH_TESTS ]; then npm run unittest; fi",
     "start-log": "node ./tools/start.js --user-data-dir-name=brave-development --enable-logging=stderr --v=1 --enable-extension-activity-logging --enable-sandbox-logging --enable-dcheck",
     "start": "node ./tools/start.js --user-data-dir-name=brave-development --enable-logging --v=0 --enable-extension-activity-logging --enable-sandbox-logging --enable-dcheck",
